requirements.txt
numpy~=1.19.5
torch~=1.9.1
torchvision~=0.10.1
Pillow~=8.4.0
opencv-python~=4.5.3.56
sklearn~=0.0
scikit-learn~=0.22.2
matplotlib~=3.3.4
opencv-python~=4.5.3.56
pandas~=1.1.5